pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financebuddy"
version = "0.2.0"
authors = [
  { name="Cedric Duriau", email="duriau.cedric@live.be" },
]
description = "FinanceBuddy is a tool that centralizes and parses data from diffent banks into a unified format, streamlining the process for analytics and reporting."
readme = "README.md"
license-files = { paths = ["LICENSE.md"] }
requires-python = ">=3.12"
dependencies = [
  "financebuddy-parsers>=0.1.0",
  "Babel>=2.16.0",
  "chardet>=5.2.0",
  "pycountry>=24.6.0",
  "pydantic>=2.9.1",
  "python-dateutil>=2.9.0",
  "tabulate>=0.9.0",
]

[project.urls]
Homepage = "https://github.com/cedricduriau/financebuddy"
Parsers = "https://github.com/cedricduriau/financebuddy-parsers"

[project.optional-dependencies]
dev = [
  "ruff>=0.0.290",
  "pip-tools>=6.12.0",
  "pytest>=7.3.1",
  "pytest-cov>=4.0.0",
]

[project.scripts]
financebuddy-cli = "financebuddy.cli:run"

[tool.ruff]
line-length = 120
exclude = [".env"]

[tool.ruff.lint]
ignore = ["E501"]
select = ["E", "F", "I", "W"]
fixable = ["I", "F"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.coverage.paths]
source = ["financebuddy"]

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    "**/__init__.py",
    "financebuddy/cli.py",
    "financebuddy/settings.py",
]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.pytest.ini_options]
addopts = "--disable-warnings --cov=financebuddy"
